+++ src/hooks/useAccessibleFocus.ts
@@ -5,10 +5,10 @@ import { useEffect, useState } from 'react';
  * will become false if they use their mouse.
  */
 export default function useAccessibleFocus() {
-  const [isUserTabbing, setIsUserTabbing] = useState(observable.getCurrentValue());
+  const [isUserTabbing, setIsUserTabbing] = useState(observable.isUserTabbing);
 
   useEffect(() => {
-    const subscriber = () => setIsUserTabbing(observable.getCurrentValue());
+    const subscriber = () => setIsUserTabbing(observable.isUserTabbing);
     const unsubscribe = observable.subscribe(subscriber);
     return unsubscribe;
   }, []);
@@ -23,14 +23,21 @@ class AccessibleFocusObservable {
   constructor() {
     this._subscribers = [];
     this._isUserTabbing = false;
-    window.addEventListener('keydown', this._handleFirstTab);
   }
 
-  getCurrentValue() {
+  get isUserTabbing() {
     return this._isUserTabbing;
   }
 
   subscribe(subscriber: Function) {
+    // If adding first subscriber, begin listening to document.
+    if (this._subscribers.length === 0) {
+      if (this._isUserTabbing) {
+        this._tabToMouseListener();
+      } else {
+        this._mouseToTabListener();
+      }
+    }
     const exists = this._subscribers.includes(subscriber);
     if (!exists) this._subscribers.push(subscriber);
     return this._unsubscribe(subscriber);
@@ -39,7 +46,10 @@ class AccessibleFocusObservable {
   private _unsubscribe(subscriber: Function) {
     return () => {
       const index = this._subscribers.indexOf(subscriber);
-      if (index !== -1) this._subscribers.splice(index, 1);
+      if (index === -1) return;
+      this._subscribers.splice(index, 1);
+      // If no subscribers, stop listening to document.
+      if (this._subscribers.length === 0) this._removeAllListeners();
     };
   }
 
@@ -51,18 +61,29 @@ class AccessibleFocusObservable {
   private _handleFirstTab = (event: KeyboardEvent) => {
     if (event.key === 'Tab') {
       this._setIsUserTabbing(true);
-
-      window.removeEventListener('keydown', this._handleFirstTab);
-      window.addEventListener('mousedown', this._handleFirstMouse);
+      this._tabToMouseListener();
     }
   };
 
   private _handleFirstMouse = () => {
     this._setIsUserTabbing(false);
+    this._mouseToTabListener();
+  };
 
+  private _tabToMouseListener() {
+    window.removeEventListener('keydown', this._handleFirstTab);
+    window.addEventListener('mousedown', this._handleFirstMouse);
+  }
+
+  private _mouseToTabListener() {
     window.removeEventListener('mousedown', this._handleFirstMouse);
     window.addEventListener('keydown', this._handleFirstTab);
-  };
+  }
+
+  private _removeAllListeners() {
+    window.removeEventListener('mousedown', this._handleFirstMouse);
+    window.addEventListener('keydown', this._handleFirstTab);
+  }
 }
 
 const observable = new AccessibleFocusObservable();