+++ src/webviewer/getRotatedDocument.ts
@@ -1,12 +1,14 @@
 import { FuturableOrLazy, futureableOrLazyToFuturable } from '../data/futurable';
 
-async function getRotatedDocument(documentObj: FuturableOrLazy<CoreControls.Document>) {
+async function getRotatedDocument(documentObj: FuturableOrLazy<CoreControls.Document>, counterclockwise?: boolean) {
   const coreControls = window.CoreControls;
   const fetchedDocument = await futureableOrLazyToFuturable(documentObj);
 
   const pageNumbers = Array.from({ length: fetchedDocument.getPageCount() }, (_v, k) => k + 1);
 
-  await fetchedDocument.rotatePages(pageNumbers, coreControls.PageRotation.e_90);
+  const rotation = counterclockwise ? coreControls.PageRotation.e_270 : coreControls.PageRotation.e_90;
+
+  await fetchedDocument.rotatePages(pageNumbers, rotation);
   return fetchedDocument;
 }